requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
Markdown==3.5.1
Pygments==2.17.2
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3